+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+###0.2.6 / 2014-10-17
+* fix http listener prefix handling
+
+###0.2.5 / 2014-10-12
+* JSON metrics refactor
+* remote metrics 
+
 ###0.2.4 / 2014-10-07
 * JSON version
 * added environment